# siron-activate-main 2/apis/v1/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Informatique-MTF S.A.
  description: Provide functionality for feature licensing management
  title: Siron Activate REST API
  version: v1.4.0
servers:
  - url: https://siron-activate.imtf-group.com
    description: Production server (uses live data)
  - url: https://licensing-service.vmo-dev-licensing.dev.imtf-group.com
    description: Sandbox server (uses test data)
...
security:
  - openId:
      - user
      - license-issuer
      - license-user
      - admin
paths:
...
  /licensing/api/v1/licenses:
    post:
      operationId: createLicense
      tags:
        - licenses
      summary: Create a new license for a capability
      security:
        - openId:
            - license-issuer
            - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/License'
            examples:
              Static example:
                value:
                  productName: My Awesome App
                  customerId: johnDoe1234
                  customerName: John Doe
                  creationDate: "2023-02-15T14:30:00.000Z"
                  expirationDate: "2025-03-01"

        description: The license to create
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/License'
              examples:
                Static example:
                  value:
                    oid: c3bf8ae1-f0b5-4535-8805-7c41909855c7
                    productName: My Awesome App
                    customerId: johnDoe1234
                    customerName: John Doe
                    creationDate: "2023-02-15T14:30:00.000Z"
                    expirationDate: "2025-03-01"

          description: The create license
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error when validating the request
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL not found

    get:
      operationId: getLicenses
      tags:
        - licenses
      summary: Get all licenses registered in the system
      security:
        - openId:
            - user
            - license-issuer
            - admin
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/License'
              examples:
                Static example:
                  value:
                    - oid: 590a6155-706d-47f4-9442-39da913643a4
                      productName: My Awesome App
                      customerId: johnDoe1234
                      customerName: John Doe
                      creationDate: "2023-02-15T14:30:00.000Z"
                      expirationDate: "2025-03-01"
                    - oid: 23456789-0123-4567-8901-234567890123
                      productName: My Other App
                      customerId: janeDoe9876
                      customerName: Jane Doe
                      creationDate: "2024-06-15T10:30:00.000Z"
                      expirationDate: "2026-09-01"

          description: A list of all licenses registered in the system
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL not found
  /licensing/api/v1/licenses/{oid}:
    get:
      operationId: getLicense
      tags:
        - licenses
      summary: Get a license by its oid
      security:
        - openId:
            - user
            - license-issuer
            - admin
      parameters:
        - in: path
          name: oid
          schema:
            type: string
          required: true
          description: the oid of the license to load
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/License'
              examples:
                Static example:
                  value:
                    oid: ec0acfcd-b4dd-4ce9-af19-34248d8769a1
                    productName: My Awesome App
                    customerId: johnDoe1234
                    customerName: John Doe
                    creationDate: "2023-02-15T14:30:00.000Z"
                    expirationDate: "2025-03-01"

          description: The license corresponding to the oid passed as parameter
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL or license oid not found
    delete:
      operationId: deleteLicense
      tags:
        - licenses
      summary: Delete a license key. Once a license has been signed, it can not be deleted anymore.
      security:
        - openId:
            - license-issuer
            - admin
      parameters:
        - in: path
          name: oid
          schema:
            type: string
          required: true
          description: the oid of the license to delete
      responses:
        "204":
          description: License successfully deleted
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized, or license has been signed and therefore can not be deleted anymore.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL or license oid not found
  /licensing/api/v1/licenses/{oid}/sign:
    post:
      operationId: signLicense
      tags:
        - licenses
      summary: Sign a license given by its oid
      security:
        - openId:
            - license-user
            - license-issuer
            - admin
      parameters:
        - in: path
          name: oid
          schema:
            type: string
          required: true
          description: the oid of the license to sign
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedLicenseString'
              example:
                value: eyJtZXRhZGF0YSI6eyJjdXN0b21lcklkIjoiaW10Zi1pbnRlcm5hbC1kZXYiLCJjdXN0b21lck5hbWUiOiJJTVRGIChpbnRlcm5hbCBkZXYpIiwiY3JlYXRpb25EYXRlIjoiMjAyNC0wOS0yMyIsImV4cGlyYXRpb25EYXRlIjoiMjAyNS0wMS0xMCIsInByb2R1Y3ROYW1lIjoiYWNtIiwic2NoZW1hVmVyc2lvbiI6MX0sInNpZ25hdHVyZSI6Ik52c2JaV0lqTE9DSVk0M3VacFFkQmorNWlSaVp0Y1JraVFjRDdEWjZiVnFxTzdvZklpYlo4clBYV2dPTm9Bcm5iWmgyd3lldXZOOUF0cUVXaWRiaE0wdmlocjRoMTEyeHJVY1FaNDZEQzNCbG0vUEwydXhjamhzNmR6ZXl3Q0JuZHEwMnlPV2FPWUhnL0trdjIzT1E3ZlVqSVAvU3lEWnFYd1hYRHdIckREbGduRVZwa2JyQ1lz4NWVWSHZrVG1ZVFV3Sk9CWGJmK3crR0RFMFV0R0hYSk1ZUFVjc05zaWVnbjIzdEY2Y0wyTWFvbDZPVENEVGFrSElPRlF6TzRQVEpDYzRMM2JQZnNMVk9ORitaV1o1NmlTTi9hUG9TeGdRdHpObmgyRXYrWS9veVhPdElYdE8xZkw5d0V6S2k0NDVsd0dIcTUwVGhyZm5HQkEwbTFPdz09In0
          description: The signed license encoded as base64
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL or license oid not found
...
components:
  securitySchemes:
    openId:
      type: openIdConnect
      openIdConnectUrl: https://keycloak.imtf-group.com/realms/imtf/.well-known/openid-configuration
      description: oidc connected to imtf internal keycloak
  schemas:
...
    License:
      type: object
      description: the license definition
      properties:
        oid:
          type: string
          description: the technical uuid assigned to the license
        capabilities:
          type: array
          description: a list of capability names for which the license has been created.
          items:
            type: string
        productName:
          type: string
          description: the name of the product for which the license has been created, it needs to match an already registered capability.
        customerId:
          type: string
          description: a string id representing the customer for which the license has been created.
        customerName:
          type: string
          description: the display name of the customer for which the license has been created. Usually, the value of this field is displayed in the capability using the license.
        creationDate:
          type: string
          format: date-time
          description: the creation date of the license
        expirationDate:
          type: string
          format: date
          description: the expiration date of the license
        signatureDate:
          type: string
          format: date-time
          description: the date of the first signature of the license (deprecated since 1.1.0)
          deprecated: true
        signedBy:
          type: string
          description: the name of the user who first signed the license (deprecated since 1.1.0)
          deprecated: true
        contactEmail:
          type: string
          format: email
          description: contact information, email or name, about the owner of this license
      required:
        - productName
        - customerId
        - customerName
        - creationDate
...
    SignedLicenseString:
      type: object
      description: a signed license
      properties:
        value:
          type: string
          description: base64 hash of the signed license
...
    ErrorResponse:
      type: object
      description: Error is used to return the error details inside error response
        object
      properties:
        code:
          type: string
          description: Unique identifier of the error type
        message:
          type: string
          description: Simple message describing the error
        timestamp:
          type: string
          format: date-time
          description: Exact time of the error, UTC based with ISO-8601 format
        spanId:
          type: string
          description: If available, the OpenTracing span Id
        traceId:
          type: string
          description: If available, the OpenTracing trace Id
      required:
        - code
        - message
        - timestamp
...